Create new CalculatorModuleList class
