+++ src/main/java/seedu/kolinux/module/ModuleList.java
@@ -61,30 +61,6 @@ public class ModuleList {
     }
 
     /**
-     * Store a module in myModules when only the modular credit and its grade are given.
-     * This method is only used CapCalculator object.
-     * 
-     * @param mc Modular credit which will be stored
-     * @param grade The corresponding grade to be stored
-     */
-    public void storeModuleMcGrade(int mc, String grade) {
-        ModuleDetails mod = new ModuleDetails(mc, grade);
-        myModules.add(mod);
-    }
-
-    /**
-     * Store a module in myModules when only the module code and its grade are needed from the module.
-     * This method is only used CapCalculator object.
-     *
-     * @param code Module code which will be stored
-     * @param grade The corresponding grade to be stored
-     */
-    public void storeModuleCodeGrade(String code, String grade) {
-        ModuleDetails mod = new ModuleDetails(code, grade);
-        myModules.add(mod);
-    }
-
-    /**
      * Stores the moduleDetails corresponding to a given module code in the myModules list.
      *
      * @param code Module code whose details will be stored
